Client e gelmeden önceki halidir

+++ Core/ETicaret.Domain/Entities/Product.cs
@@ -11,7 +11,8 @@ namespace ETicaretAPI.Domain.Entities
     {
         public string Name { get; set; }
         public int Stock { get; set; }
-        public long Price { get; set; }
+        public float Price { get; set; }
+
